add statements for basic_def
I don't want to write a proper parser, and we have one already, which
is why the expressions are a little hard to read.
It's not so important how they are written, since they will be checked
along with everything else; they just have to match the expectation of later
theorems.