bug hunting, general testing
One relevant change in this commit is that `TyVarId` identifiers are
now global, making them more similar to the original ocaml
implementation. I really wanted to use ordered lists and numbered local
variables here, but unfortunately the proof rules are really unhygienic:

* The Fetch rule does not require that you substitute into all type
  variables; you can just leave some unsubstituted and then it keeps the
  "original type variable" (which in our setting means the variable has
  its name copied from the old one, possibly matching an existing type
  variable if the name has already been used).

* In the new_type_definition command, if you declare a type from a
  predicate P, it will go through the type variables in text order and
  then reorder them to be asciibetical (or whatever (<=) on string means
  in ocaml). So we can't even use global identifiers in place of
  strings. (Well, we do anyway, and just look up the strings for this
  particular use case.)

So as usual, the proof rules are baking in some really unnecessary ocaml
details, and we have to pretend to be ocaml to check proofs. We'll clean
this up during the export phase instead of import.