checked flyspeck! clean up output, report timing
There appears to be a bad proof step in the official flyspeck replay,
in which theorem `Nthm_1.prf` exists in both `BaseSystem` and `Local5`
modules (and it's not a benign redefinition). It is skipped over by the
HOL Zero replayer because it checks if the theorem already exists
before reading the statement. So I'm mimicking that behavior now and it
makes it to the end.